Fixed testing so that user none is no longer null
(This is so it works better with DB)

PROBLEM:

One test is now failing. I believe this is due to DB adding a game to a
user when an invite is accepted, but we already do it too.

Potential solution is to comment out the following lines from
Backend/Invite.java since it seems like Torry's code does this too: 

		userTo.addCurrentGame(gme);
		userFrom.addCurrentGame(gme);
		
Lines are contained in acceptInvite() method